[CPU] Refactor IQ3_XXS dequant kernels into shared nodes/kernels helper (M7.c / B-1)
Extracts the IQ3_XXS dequantize-on-the-fly + fused dot product kernel that
previously lived inline inside nodes/iq3_xxs_linear.cpp into a dedicated public
header pair under nodes/kernels/ so it can be reused by other consumers in the
intel_cpu plugin (in particular the upcoming CompressedConstantFCExecutor).

New files:
  * src/plugins/intel_cpu/src/nodes/kernels/iq3_xxs_kernels.hpp
  * src/plugins/intel_cpu/src/nodes/kernels/iq3_xxs_kernels.cpp

Public API in namespace ov::intel_cpu::kernels::iq3_xxs:

  void decode_iq3_xxs_row(const uint8_t* w_row,
                          float*         out,
                          size_t         blocks_per_row);

  void iq3_xxs_fc(const float*   act,
                  const uint8_t* weight_blob,
                  float*         out,
                  size_t         rows,
                  size_t         K,
                  size_t         N,
                  size_t         nthreads_hint = 0);

Bit-exact contract — the decode logic is a byte-for-byte translation of
ggml-quants.c::dequantize_row_iq3_xxs. The 256-entry kIq3xxsGrid and 128-entry
kSignsIq2xs tables are copied verbatim from ggml-common.h. The fp16_to_f32
helper is a pure scalar IEEE-754 implementation that does NOT depend on the
host CPU's f16 support, so output is reproducible across builds.

iq3_xxs_fc is internally parallelized over N using std::thread (this build
uses THREADING=SEQ which makes ov::parallel_for serial). Each thread decodes
its slab of N rows once into a per-thread K-float buffer and reuses it across
all activation rows, so weight memory never grows.

nodes/iq3_xxs_linear.cpp shrinks from 299 to 113 lines: removed inline tables,
inline fp16_to_f32, inline decode_iq3_xxs_row, and the inline worker /
std::thread pool. execute() now just adapts MemoryArgs to plain pointer/size
arguments and calls kernels::iq3_xxs::iq3_xxs_fc. The OV_CPU_IQ3XXS_TRACE
env-var diagnostic is preserved.

Validation: cc_dual_route_probe Route A produces
[224.723404, 35.0891609, 129.633942, -240.220047] for the canonical test
blob — byte-identical to the pre-refactor output. ApiValidator OneCore
compliance check passes after the refactor.
